Created parser header
Includes the definition of TreeNode as well as ASTNode. ASTNode contains an expression TreeNode, which holds a mini AST of for basic expressions. The ASTNode's purpose is to give more context of the expression including the statements type as well as acting as a container for full expressions including return statements for functions, holding an entire body of expressions, a possible data type or identifier for functions, and finally any arguments it was given for conditionals or functions.